argentum_ui::components::primitives

Module kbd

Structs

KbdGroupProps

A row of kbd keys making up one shortcut.

KbdGroupPropsBuilder

Typestate builder for KbdGroupProps, created by KbdGroupProps::builder(). build() becomes available once every required property has been set.

KbdProps

A keyboard key component: a <kbd> drawn as a key cap.

KbdPropsBuilder

Typestate builder for KbdProps, created by KbdProps::builder(). build() becomes available once every required property has been set.

kbd

A keyboard key component: a <kbd> drawn as a key cap.

kbd_group

A row of kbd keys making up one shortcut.